+++ src/HeboTech.ATLib.Modems/Cinterion/MC55i.cs
@@ -1,5 +1,5 @@
 ﻿using HeboTech.ATLib.Modems.Generic;
-using HeboTech.ATLib.Parsers;
+using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Sys;
 using System;
