+++ src/HeboTech.ATLib.Modems/Generic/ModemBase.cs
@@ -2,7 +2,7 @@
 using HeboTech.ATLib.Extensions;
 using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Numbering;
-using HeboTech.ATLib.Parsers;
+using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Storage;
 using HeboTech.ATLib.Sys;
 using System;
