+++ src/HeboTech.ATLib.Modems/TP-LINK/MA260.cs
@@ -1,5 +1,5 @@
 ﻿using HeboTech.ATLib.Modems.Generic;
-using HeboTech.ATLib.Parsers;
+using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Messaging;
 using System.Collections.Generic;
 using System.Threading.Tasks;
