+++ src/HeboTech.ATLib.Tests/Sys/BatteryChargeStatusTests.cs
@@ -2,7 +2,7 @@
 using HeboTech.ATLib.Sys;
 using Xunit;
 
-namespace HeboTech.ATLib.Tests.System
+namespace HeboTech.ATLib.Tests.Sys
 {
     public class BatteryChargeStatusTests
     {
