+++ src/HeboTech.ATLib/IModem.cs
@@ -1,7 +1,7 @@
 ﻿using HeboTech.ATLib.Events;
 using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Numbering;
-using HeboTech.ATLib.Parsers;
+using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Storage;
 using HeboTech.ATLib.Sys;
 using System;
