+++ src/HeboTech.ATLib/Parsing/AtChannel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace HeboTech.ATLib.Parsers
+namespace HeboTech.ATLib.Parsing
 {
     public class AtChannel : IAtChannel, IDisposable
     {
