+++ src/HeboTech.ATLib/Parsing/AtCommandType.cs
@@ -1,4 +1,4 @@
-﻿namespace HeboTech.ATLib.Parsers
+﻿namespace HeboTech.ATLib.Parsing
 {
     internal enum AtCommandType
     {
