+++ src/HeboTech.ATLib/Parsing/AtWriter.cs
@@ -4,7 +4,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace HeboTech.ATLib.Parsers
+namespace HeboTech.ATLib.Parsing
 {
     internal class AtWriter : IAtWriter, IDisposable
     {
