+++ src/HeboTech.ATLib/Parsing/IAtChannel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace HeboTech.ATLib.Parsers
+namespace HeboTech.ATLib.Parsing
 {
     public interface IAtChannel
     {
