+++ src/HeboTech.ATLib/Parsing/ModemResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace HeboTech.ATLib.Parsers
+﻿namespace HeboTech.ATLib.Parsing
 {
     public class ModemResponse
     {
